+++ public/pages/CreateIndexTemplate/components/DefineTemplate/index.ts
@@ -3,5 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import DefineTemplate from "./DefineTemplate";
+export { default as OverviewTemplate } from "./OverviewTemplate";
 
 export default DefineTemplate;
